Extract Git utils into module
